requirements.txt
python-binance==0.7.5
python-dotenv==0.15.0
sentry-sdk==0.19.5
freezegun==1.0.0
ring==0.10.1
requests==2.26.0
enlighten==1.7.1
pandas-ta==0.2.23b0
ta==0.10.2
mplfinance==0.12.9b7
appdirs~=1.4.4